refactor(product) : 상품 Bounded Context로 패키지 구성
