+++ src/main/java/kitchenpos/product/domain/JdbcTemplateProductDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.product.domain;
 
-import kitchenpos.model.Product;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
